refactor(admin): split admin-panel into per-tab modules; integrate Users tab
The 1781-line admin-panel.ts violated §II.3 of the constitution
(target <500, hard ceiling 750 lines). This refactor extracts each
tab's rendering and state into its own module under
`src/components/admin-tabs/`, leaving admin-panel.ts as a 230-line
shell that owns the year selector, tab strip, and shared data
caches (teams, season, name suggestions).

The Users tab — previously rendered as a sibling Custom Element via
main.ts._applyAdminViewState — is now a 4th tab inside admin-panel,
hosting <admin-users-panel> in its tab body. The component is
self-gating (mounts/fetches only when role becomes owner|admin), so
it works as-is inside the tab.

Module breakdown:
- admin-panel.ts (230) — shell: tab strip, year selector,
  shared caches, tab module lifecycle
- admin-tabs/types.ts (54) — AdminTab + AdminTabContext interfaces
- admin-tabs/admin-shared.ts (224) — SVG icons, buildOptions,
  parseLocalDate, toInputDate, showConfirmDialog,
  attachAutocomplete, setStatus
- admin-tabs/team-management-tab.ts (331) — team list with
  inline edit
- admin-tabs/roster-modal.ts (305) — per-team shooter editor
- admin-tabs/score-entry-tab.ts (583) — week/team picker,
  date override, save/publish flow (single tab UI; locked-state
  semantics tightly couple date override to score entry)
- admin-tabs/announcements-tab.ts (276) — banner + announcements

main.ts: drop sibling <admin-users-panel> from the lazy-mount
innerHTML — it now mounts as a child of <admin-panel> via the
Users tab body.

Drops dead method `_confirmRemoveShooterRow` (defined but never
referenced anywhere in the codebase).

Constitutional compliance:
- §II.3: all new files within target/ceiling; admin-panel.ts is
  230 lines (under <500 success criterion)
- §III.5: TypeScript strict; no var, no inline handlers
- §IV.2: no Firestore-direct components; admin-users-panel still
  routes through admin-user-service

Tests: typecheck clean; 178/178 unit tests pass. Rules/Functions
tests not run (emulator port occupied) but rules + functions code
unchanged — no test paths touched.
